Adjusted and did small reorganizations in message parsing logic. The final results appear to function as intended, so the parser appears to be finished.
Results will contain a Type that helps identify which struct the result has, along with the keys and values processed.

The next step is to create V2 of the processing and printing of the results that get parsed from chat events, in the reporting.lua which will deal with accumulation and rendering of results.