Cleaned up the `LootRoll` related messages and formatting handling.
Moved the GetChatFrame into the settings, and made a db proxy so we don't have to consider the SV event, we can either load the defaults (if we run it that early), or the actual SV data.

If Debounce is 0 we instantly flush and never wait.

We ensure to check for `db.EnabledGroups` being `false` and not just `nil` when considering it disabled.