Added a deparate table for custom debounce timers, specifically used for loot roll related event that should occur at once.
The output handler and buffer can now flush and clear a specific group when required.

Minor change so the standard SV table doesn't have pre-filled `true` in EnabledGroups. Instead, we specifically check for `false` when we want to know if it's actually set on purpose (instead of falsy/nil).

Implemented the tooltips system to anchor tooltips on top the chat frame when hovering supported hyperlinks.

WIP system for formatter output type to allow future grouping when printing, like if we wish to "sum up all lines" related to the same person, for now we just print data based on match-types, so items and currencies are in two separate lines even if it's the same person.