The filters export `EvaluateFilters` that is used to evaluate the `db.Filters`.
Made it possible to setup both single rules, or a rule set with `and`/`or` logic, to further nest additional rules. This should allow as much flexibility as we might need.

The formatter colors quest items in red border, for the sake of visibility.

Fixed issue with numbers not being formatted properly when being printed.