Fixed issue with `SumByKey` it didn't consider itself as `1` count. So if the value is nil we just add 1 as standard behavior. The method supports a skip-flag for this behavior if needed.
Updated `ProcessTokens` to cleanup potential malformed links. When a link is expected, we try to extract it from the string.

Sometimes the string matching will contain additional syntax, that in most cases is fine, but Currency does not like additional data when working with those type of links.

Try to extract the complete color and link data, fallback to only link data, and if any match use the match as the value as it is the link we're looking for.

Minor docs and type checking adjustments.